Added freezing functions as their files for DUET and NMF MFCC.
